requirements.txt
pyarrow
snowflake-snowpark-python
snowflake-ml-python        # PyTorchDistributor, Tuner, MLJob
